+++ src/source.cu
@@ -186,7 +186,7 @@ int run(int argc, char* argv[])
         }
 
         std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now();
-        time = std::chrono::duration_cast<std::chrono::microseconds>(end-begin).count();
+        time = static_cast<float>(std::chrono::duration_cast<std::chrono::microseconds>(end-begin).count());
     }
     printf("\n");
